requirements.txt
PyQt6
photoshop-python-api==0.19.0 